+++ src/features/auth/container/login-form/LoginForm.tsx
@@ -6,26 +6,9 @@ import SubmitButton from '../../components/submit-button/SubmitButton';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { loginFormSchema, LoginFormData } from '../../types/loginFormSchema';
-import { login } from '../../api/auth';
-import { useRouter, useSearchParams } from 'next/navigation';
-import {
-  AUTH_ERROR_MESSAGES,
-  LOGIN_FORM_PLACEHOLDERS,
-} from '../../constants/messages';
-
-type LoginErrorCode = 'USER_NOT_FOUND' | 'INVALID_CREDENTIALS' | 'SERVER_ERROR';
-
-interface LoginError {
-  response: {
-    data: {
-      code: LoginErrorCode;
-    };
-  };
-}
+import { useLoginSubmit } from '../../hooks/useLoginSubmit';
 
 export default function LoginForm() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
   const {
     register,
     handleSubmit,
@@ -37,36 +20,7 @@ export default function LoginForm() {
     mode: 'onChange',
   });
 
-  const onSubmit = async (data: LoginFormData) => {
-    try {
-      const response = await login(data);
-      console.log('로그인 성공:', response);
-      reset();
-
-      const returnUrl = searchParams.get('returnUrl') || '/';
-      router.replace(returnUrl);
-    } catch (error) {
-      const { code } = (error as LoginError).response.data;
-
-      switch (code) {
-        case 'USER_NOT_FOUND':
-          setError('email', {
-            type: 'manual',
-            message: AUTH_ERROR_MESSAGES.USER_NOT_FOUND,
-          });
-          break;
-        case 'INVALID_CREDENTIALS':
-          setError('password', {
-            type: 'manual',
-            message: AUTH_ERROR_MESSAGES.INVALID_CREDENTIALS,
-          });
-          break;
-        case 'SERVER_ERROR':
-          console.error(AUTH_ERROR_MESSAGES.SERVER_ERROR);
-          break;
-      }
-    }
-  };
+  const onSubmit = useLoginSubmit(setError, reset);
 
   return (
     <div className="flex h-auto w-full flex-col items-center justify-center rounded-3xl bg-white px-4 py-8 sm:h-[422px] sm:w-[510px] sm:px-14">
@@ -80,7 +34,7 @@ export default function LoginForm() {
           <FormField
             label="아이디"
             type="text"
-            placeholder={LOGIN_FORM_PLACEHOLDERS.EMAIL}
+            placeholder="이메일"
             id="email"
             register={register('email')}
             error={errors.email?.message}
@@ -88,7 +42,7 @@ export default function LoginForm() {
           <FormField
             label="비밀번호"
             type="password"
-            placeholder={LOGIN_FORM_PLACEHOLDERS.PASSWORD}
+            placeholder="비밀번호"
             id="password"
             register={register('password')}
             error={errors.password?.message}
